% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_taxonomy.R
\name{add_taxonomy}
\alias{add_taxonomy}
\title{Retrieve and add taxonomic information to a \code{fwspp} object}
\usage{
add_taxonomy(fwspp, taxonomy = NULL)
}
\arguments{
\item{fwspp}{an \code{fwspp} object returned by \code{\link{fws_occ}}}

\item{taxonomy}{a \code{data.frame} returned by
\code{\link{retrieve_taxonomy}}; this is meant primarily for internal
\code{fws_occ} use}
}
\value{
a \code{\link{fwspp}} object with added (or updated) basic
 taxonomic information
}
\description{
Use this function to add or update basic taxonomic information to an
 \code{fwspp} object; see \code{\link{retrieve_taxonomy}} for
 a description of the information.
}
\examples{
\dontrun{
ml <- find_fws("longleaf")
ml <- fws_occ(fws = ml)
ml <- add_taxonomy(ml)
}
}
