% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{buffer_prop}
\alias{buffer_prop}
\title{Generate a spatial buffer around a USFWS property}
\usage{
buffer_prop(prop, buffer)
}
\arguments{
\item{prop}{a FWS property boundary returned by \code{\link{find_fws}}}

\item{buffer}{a numeric scalar indicating the buffer radius in meters}
}
\value{
A FWS property boundary buffered by the buffer radius
}
\description{
Generate a spatial buffer around a USFWS property
}
