% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{check_cadastral}
\alias{check_cadastral}
\title{Check that necessary USFWS cadastral geodatabase is installed}
\usage{
check_cadastral()
}
\description{
Check that necessary USFWS cadastral geodatabase is installed
}
