% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_properties.R
\name{dice_polys}
\alias{dice_polys}
\title{Do the actual splitting of large properties into more manageable units}
\usage{
dice_polys(polys, max_area)
}
\arguments{
\item{polys}{\code{\link[sf]{sf}} for a single USFWS property}

\item{max_area}{numeric scalar indicating the approximate maximum area
(in m^2) to allow when splitting \code{ref}. Most will be smaller, often
considerably so, given the irregular boundaries of most properties}
}
\description{
Do the actual splitting of large properties into more manageable units
}
