% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{dice_prop}
\alias{dice_prop}
\title{Split large FWS property into smaller more manageable units}
\usage{
dice_prop(prop)
}
\arguments{
\item{prop}{a FWS property boundary returned by \code{\link{find_fws}}}
}
\value{
A FWS property split into smaller units
}
\description{
Split large FWS property into smaller more manageable units
}
