% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{ensure_multipolygons}
\alias{ensure_multipolygons}
\title{Ensure the the cadastral data is a multipolygon object.}
\usage{
ensure_multipolygons(prop)
}
\arguments{
\item{prop}{a FWS property boundary returned by \code{\link{find_fws}}}
}
\value{
a multipolygon sf object
}
\description{
Addresses https://github.com/r-spatial/sf/issues/427#issuecomment-874476218
}
