% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwspp_combine.R
\name{fwspp_combine}
\alias{fwspp_combine}
\title{Consolidate two or more \code{fwspp} objects}
\usage{
fwspp_combine(...)
}
\arguments{
\item{...}{two or more \code{fwspp} objects returned by \code{\link{fws_occ}}}
}
\value{
a \code{\link{fwspp}} object
}
\description{
Use this function to combine two or more \code{fwspp} objects into a single
  \code{fwspp} object. Attributes are preserved. Only works if \code{fwspp}
  objects were generated using the same \code{boundary}, \code{scrubbing},
  \code{buffer}, and \code{taxonomy} settings (see \code{\link{fws_occ}}.
  The earliest query date/time in the combined \code{fwspp} objects is
  assigned to the combined object.
}
\examples{
\dontrun{
ml <- find_fws("longleaf")
ml <- fws_occ(fws = ml)
mb <- find_fws("mountain bog")
mb <- fws_occ(fws = mb)
mlb <- fwspp_combine(ml, mb)
}
}
