% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwspp_review.R
\name{fwspp_review}
\alias{fwspp_review}
\title{Export species occurrence records in an \code{fwspp} object to separate
 files for review}
\usage{
fwspp_review(fwspp, dir = "./fwspp_review", overwrite = FALSE,
  verbose = TRUE)
}
\arguments{
\item{fwspp}{a \code{fwspp} object returned by \code{\link{fws_occ}},
**with taxonomy information added** (see Details). An output file will be created in \code{dir}
for each property in \code{fwspp}.}

\item{dir}{a non-empty character scalar of path to store an output file for
each USFWS property in \code{fwspp}. The default is to use or create a
\code{fwspp_review} directory in the current working directory.}

\item{overwrite}{logical (default \code{FALSE}); overwrite existing files
with the same name?}

\item{verbose}{logical (default \code{TRUE}); provide detailed messaging during
export process?}
}
\value{
\code{NULL}; Exports individual Excel file(s) to \code{dir} for
 the review of species occurrence observations
}
\description{
Use this function to export species occurrence records into spreadsheets
 for review prior to parsing for inclusing into the FWSpecies database.
}
\details{
This function **requires** that the \code{fwspp} object have taxonomy
 added (either via the \code{taxonomy = TRUE} argument during the call
 to \code{\link{fws_occ}} or subsequently via \code{\link{add_taxonomy}}).
 Reviewing records without additional taxonomic information (such as common
 name and general taxonomic groupings) is likely to be frustrating.
}
\examples{
\dontrun{
lowcountry <- find_fws(c("romain", "santee", "ace basin", "waccamaw"))
lc <- fws_occ(fws = lowcountry)
review_fwspp(lc)
}
}
