% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_grbio.R
\name{get_grbio}
\alias{get_grbio}
\title{Retrieve GRBIO institution codes}
\usage{
get_grbio()
}
\description{
Retrieve institution codes for contributors to the Global Registry of
Biodiversity Repositories (\url{http://grbio.org/})
}
