% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{get_UTM_zone}
\alias{get_UTM_zone}
\title{Get the UTM zone for a given longitude}
\usage{
get_UTM_zone(lon)
}
\arguments{
\item{lon}{numeric scalar of longitude in decimal degrees}
}
\value{
numeric value that is the UTM zone
}
\description{
Get the UTM zone for a given longitude
}
