% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{get_wkt}
\alias{get_wkt}
\title{Get WKT (Well-Known Text) representation of an sf multipolygon object}
\usage{
get_wkt(prop)
}
\arguments{
\item{prop}{a FWS property boundary returned by \code{\link{find_fws}}}
}
\value{
A FWS property boundary in WKT format, simplifies to a bounding box if it's complex
}
\description{
Get WKT (Well-Known Text) representation of an sf multipolygon object
}
