% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_helpers.R
\name{has_taxonomy}
\alias{has_taxonomy}
\title{Check if \code{fwspp} object has taxonomic information attached}
\usage{
has_taxonomy(fwspp)
}
\arguments{
\item{fwspp}{an \code{fwspp} object returned by \code{\link{fws_occ}}}
}
\description{
Check if \code{fwspp} object has taxonomic information attached
}
