% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{ll_ratio}
\alias{ll_ratio}
\title{Calculate approximate ratio of latitude:longitude distance}
\usage{
ll_ratio(lat)
}
\arguments{
\item{lat}{numeric scalar of latitude in decimal degrees}
}
\description{
Used to try and generate relatively square polygons when chopping
 properties
}
