% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_gets.R
\name{manage_gets}
\alias{manage_gets}
\title{Manage the retrieval, cleaning, and error handling of
occurrence records for a given \code{\link[sf]{sf}} object}
\usage{
manage_gets(poly, grbio, timeout)
}
\arguments{
\item{poly}{\code{\link[sf]{sf}} object; must contain only one POLYGON
or MULTIPOLYGON geometry}

\item{grbio}{\code{data.frame} created by \code{\link{get_grbio}}}

\item{timeout}{integer indicating time, in seconds, to allow for HTTP requests to
process.}
}
\description{
Manage the retrieval, cleaning, and error handling of
occurrence records for a given \code{\link[sf]{sf}} object
}
