% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_helpers.R
\name{nps_taxonomy}
\alias{nps_taxonomy}
\title{Retrieve raw NPS taxonomic information matching scientific name query}
\usage{
nps_taxonomy(sci_name)
}
\arguments{
\item{sci_name}{scientific name character scalar (case-insensitive) for
which to retrieve basic taxonomic information. Unlike
\code{\link{retrieve_taxonomy}}, matches are not restricted to the
species level; that is, subspecies may be returned.}
}
\value{
a \code{data.frame} of basic taxonomic infomration or \code{NULL}
 if there are no matching records
}
\description{
Queries National Park Service taxonomy records by scientific name,
 returning all records that match the input \code{sci_name} parameter
 (including partial matches; see Details). Returns basic taxonomic
 information, if available, including ITIS Taxonomic Serial Number
 (\code{tsn}), the \code{taxon_code} assigned by the National Park
 Service (and used also by the US Fish & Wildlife Service),
 \code{common_name}(s), a generic taxon group, whether the taxon is
 valid according to ITIS and, if not, the accepted \code{taxon_code}
 if a match is found.
}
\details{
It is somewhat unclear how the NPS records are filtered. The best guess
 is that it parses the unique words in the query string (\code{sci_name})
 and returns all records that match that set of unique words. The order
 of query words is not strictly enforced. See examples.
}
\examples{
\dontrun{
nps_taxonomy("GULo gulo")
nps_taxonomy("Lampropeltis holbrooki")
nps_taxonomy("holBRooki lampropeltis")
}
}
