% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_properties.R
\name{poly_bb_area}
\alias{poly_bb_area}
\title{Calculate bounding box area(s) for \code{sf} object}
\usage{
poly_bb_area(polys)
}
\arguments{
\item{polys}{\code{\link[sf]{sf}} object with POLYGON or MULTIPOLYGON geometry}
}
\description{
Calculate bounding box area(s) for \code{sf} object
}
