% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cadastral.R
\name{prep_cadastral}
\alias{prep_cadastral}
\title{Prepare USFWS cadastral spatial data}
\usage{
prep_cadastral()
}
\description{
Converts USFWS cadastral geodatabase into two \code{\link[sf]{sf}} and
 stores them for faster loading on subsequent use by
 \code{link{prep_properties}}. Exported but not typically called
 directly by user unless \code{\link{install_fws_cadastral}} fails.
}
