% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_properties.R
\name{prep_properties}
\alias{prep_properties}
\title{Filter USFWS properties for query}
\usage{
prep_properties(prop_df, bnd, verbose)
}
\arguments{
\item{prop_df}{\code{data.frame} of organizational names (ORGNAME) and
type (RSL_TYPE) of USFWS properties and their associated USFWS region
(FWSREGION). It is strongly recommended to use the results generated
by \code{\link{find_fws}} to ensure proper content and formatting.}

\item{bnd}{character scalar; which property boundary to use.  See
\code{\link{fws_occ}}}

\item{verbose}{logical; suppress messaging? See \code{\link{fws_occ}}
during species occurrence queries}
}
\description{
Filters USFWS cadastral \code{\link[sf]{sf}} data to the relevant
 properties. Exported by not typically called directly by user.
}
