% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{prop_bb}
\alias{prop_bb}
\title{Make property bounding box around a FWS property boundary}
\usage{
prop_bb(prop)
}
\arguments{
\item{prop}{a FWS property boundary returned by \code{\link{find_fws}}}
}
\value{
A bounding box for a FWS property
}
\description{
Make property bounding box around a FWS property boundary
}
