% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{prop_bb_area}
\alias{prop_bb_area}
\title{Calculate the area of a FWS property bounding box}
\usage{
prop_bb_area(prop)
}
\arguments{
\item{prop}{a FWS property boundary returned by \code{\link{find_fws}}}
}
\value{
The area of a FWS property bounding box
}
\description{
Calculate the area of a FWS property bounding box
}
