% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy_helpers.R
\name{retrieve_taxonomy}
\alias{retrieve_taxonomy}
\title{Retrieve taxonomic information for one or more scientific names}
\usage{
retrieve_taxonomy(sci_name)
}
\arguments{
\item{sci_name}{character vector (case-insensitive) of scientific names
for which to retrieve basic taxonomic information at the *species* level;
that is, subspecies (trinomials) are currently ignored}
}
\value{
a \code{data.frame} of basic taxonomic information
}
\description{
Retrieves very basic taxonomic information for a given taxa, if available,
 including Taxonomic Serial Number (\code{tsn}), the \code{taxon_code}
 used also by the US Fish & Wildlife Service), \code{common_name}(s),
 a generic taxon group, and a \code{note} if a match is not found.
}
\examples{
\dontrun{
retrieve_taxonomy(c("GULo gulo", "Lampropeltis getuLA HOLBrookI",
                    "Lampropeltis holbrooki", "Pseudemys scripta",
                    "Fakus speciesus", "Salsola iberica"))
}
}
