% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{split_at_idl}
\alias{split_at_idl}
\title{Split FWS property at International Date Line}
\usage{
split_at_idl(prop)
}
\arguments{
\item{prop}{a FWS property boundary returned by \code{\link{find_fws}}}
}
\value{
A FWS property that is split along the international date line
}
\description{
Split FWS property at International Date Line
}
