% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{split_prop}
\alias{split_prop}
\title{Split FWS properties into smaller pieces to avoid timeout errors when
querying species occurrence records from databases}
\usage{
split_prop(prop)
}
\arguments{
\item{prop}{a FWS property boundary returned by \code{\link{find_fws}}}
}
\value{
sf object of diced FWS properties
}
\description{
Split FWS properties into smaller pieces to avoid timeout errors when
querying species occurrence records from databases
}
