% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadastral_helpers.R
\name{stash_cadastral}
\alias{stash_cadastral}
\title{Store USFWS cadastral data prior to updating \code{fwspp} package}
\usage{
stash_cadastral()
}
\value{
character string of stashed cadastral files; pass this string
 to \code{\link{unstash_cadastral}} after re-installing the package
}
\description{
Regular updates of the \code{fwspp} package can be annoying because
 prepping the USFWS cadastral data consumes 10-15 minutes. Use this
 function to stash previously installed cadastral data prior to
 updating or reinstalling the \code{fwspp} package and then run
 \code{\link{unstash_cadastral}} to restore the cadastral to its
 rightful place. It's best not to load the \code{fwspp} namespace
 when stashing the cadastral data. See example.
}
\examples{
\dontrun{
# In new session
cad <- fwspp::stash_cadastral()

# Re-install fwspp package
devtools::install_github("adamdsmith/fwspp")

# Unstash the cadastral data
library(fwspp)
unstash_cadastral(cad)
}
}
