% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cadastral_helpers.R
\name{unstash_cadastral}
\alias{unstash_cadastral}
\title{Restore USFWS cadastral data after updating \code{fwspp} package}
\usage{
unstash_cadastral(stash)
}
\arguments{
\item{stash}{character vector of temporary file paths (length == 3)
generated by \code{\link{stash_cadastral}}}
}
\value{
\code{NULL}. Stashed cadastral files are restored.
}
\description{
Regular updates of the \code{fwspp} package can be annoying because
 prepping the USFWS cadastral data consumes 10-15 minutes. Use this
 function to unstash USFWS cadastral data stashed with
 \code{\link{unstash_cadastral}} prior to updating or reinstalling
 the \code{fwspp} package. See example.
}
\examples{
\dontrun{
# In new session
cad <- fwspp::stash_cadastral()

# Re-install fwspp package
devtools::install_github("USFWS/fwspp")

# Unstash the cadastral data
library(fwspp)
unstash_cadastral(cad)
}
}
