% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_helpers.R
\name{ll_ratio}
\alias{ll_ratio}
\title{Calculate approximate ratio of latitude:longitude distance	for a given
latitudinal position on the Earth}
\usage{
ll_ratio(lat)
}
\arguments{
\item{lat}{numeric scalar of latitude in decimal degrees}
}
\value{
numeric ratio of latitude:longitude distance	for a given latitudinal position
}
\description{
Calculate approximate ratio of latitude:longitude distance	for a given
latitudinal position on the Earth
}
