% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_cadastral.R
\name{prep_cadastral}
\alias{prep_cadastral}
\title{Prepare USFWS cadastral spatial data}
\usage{
prep_cadastral(refuge, bnd, verbose)
}
\arguments{
\item{refuge}{character string; USFWS properties from which to extract
species occurrence records.  See \code{\link{fw_spp}}.}

\item{bnd}{character scalar; which refuge boundary to use.  See
\code{\link{fw_spp}}}

\item{verbose}{logical; suppress messaging? See \code{\link{fw_spp}}
during species occurrence queries}
}
\description{
Converts USFWS cadastral geodatabase into \code{\link[sf]{sf}} and
 filters to relevant properties, if necessary.  Not meant to be called
 directly by user.
}
