% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_properties.R
\name{split_prop}
\alias{split_prop}
\title{If necessary, split large properties into more manageable units for queries}
\usage{
split_prop(prop, area_cutoff = 10^2)
}
\arguments{
\item{prop}{\code{\link[sf]{sf}} object for a single USFWS property}

\item{area_cutoff}{numeric scalar indicating the polygon area (in km^2) above
which to split \code{prop} into smaller, approximately \code{area_cutoff}-sized
polygons for processing}
}
\description{
Based on some preliminary timing comparisons, we split properties with one
 or more polygons occupying a bounding box larger that 100 km^2
}
